requirements.txt
Flask==3.0.0
Flask_Bootstrap==3.3.7.1
Flask_Cors==4.0.0
python-dotenv==1.0.0
